zzinzzinmak

+++ Libft/ft_split.c
@@ -6,7 +6,7 @@
 /*   By: jiyseo <jiyseo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/11 02:34:25 by jiyseo            #+#    #+#             */
-/*   Updated: 2022/01/22 03:22:57 by jiyseo           ###   ########.fr       */
+/*   Updated: 2022/01/22 03:26:48 by jiyseo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,9 +26,9 @@ char	**ft_free(char	**res)
 	return (NULL);
 }
 
-long long	ft_word_cnt(char const *str, char c)
+unsigned int	ft_word_cnt(char const *str, char c)
 {
-	long long	cnt;
+	unsigned int	cnt;
 
 	cnt = 0;
 	while (*str)
